Fix job counting bugs and dead code in node
- Tighten update_out() guard from < to <= so a departure with zero
  jobs in the node is caught immediately rather than one call late
- Remove dead double-accumulation block in update_out() that always
  added zero (last_event had just been set to the current time)
- Fix throughput denominator in to_string() to use the maximum
  last_event across all classes instead of the last class only,
  correcting the value in multi-class configurations
- Remove dead loop in to_string() that built unused nm strings
- Widen in/out counters from int to long to avoid overflow in
  long-running or high-throughput simulations

+++ src/libdes_node.hpp
@@ -273,8 +273,8 @@ class des::node : public des::object, public des::observable
 		// ids generator
 		inline static unsigned int id_gen = 0;
 		// Measures
-        vector<int> in;
-        vector<int> out;
+        vector<long> in;
+        vector<long> out;
         vector<double> usage;
         vector<double> last_event;
 		// Utility methods
